# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Đăng ký tài khoản",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Đăng nhập",
        "tags": [
          "auth"
        ]
      }
    },
    "/matches": {
      "post": {
        "operationId": "MatchingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Tạo kèo từ 1 booking đã CONFIRMED",
        "tags": [
          "matches"
        ]
      },
      "get": {
        "operationId": "MatchingController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/matches/{id}": {
      "get": {
        "operationId": "MatchingController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      },
      "patch": {
        "operationId": "MatchingController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      },
      "delete": {
        "operationId": "MatchingController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
    "/matches/{id}/join": {
      "post": {
        "operationId": "MatchingController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Xin ghép kèo",
        "tags": [
          "matches"
        ]
      }
    },
    "/venues": {
      "post": {
        "operationId": "VenueController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVenueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Tạo cụm sân",
        "tags": [
          "venues"
        ]
      },
      "get": {
        "operationId": "VenueController_findAll",
        "parameters": [
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sport",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Tìm sân theo vị trí + bộ môn",
        "tags": [
          "venues"
        ]
      }
    },
    "/venues/{id}": {
      "get": {
        "operationId": "VenueController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "venues"
        ]
      },
      "patch": {
        "operationId": "VenueController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "venues"
        ]
      },
      "delete": {
        "operationId": "VenueController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "venues"
        ]
      }
    },
    "/courts/{id}/slots": {
      "get": {
        "operationId": "CourtController_getSlots",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Danh sách ô giờ còn trống theo ngày",
        "tags": [
          "courts"
        ]
      }
    },
    "/bookings": {
      "post": {
        "operationId": "BookingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Đặt sân (⭐ race condition)",
        "tags": [
          "bookings"
        ]
      },
      "get": {
        "operationId": "BookingController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "operationId": "BookingController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bookings"
        ]
      },
      "patch": {
        "operationId": "BookingController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bookings"
        ]
      },
      "delete": {
        "operationId": "BookingController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/{id}/cancel": {
      "post": {
        "operationId": "BookingController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Hủy đơn đặt sân",
        "tags": [
          "bookings"
        ]
      }
    },
    "/merchant/revenue": {
      "get": {
        "operationId": "MerchantController_getRevenue",
        "parameters": [
          {
            "name": "range",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "year"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Thống kê doanh thu chủ sân",
        "tags": [
          "merchant"
        ]
      }
    },
    "/payments": {
      "post": {
        "operationId": "PaymentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      },
      "get": {
        "operationId": "PaymentController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{id}": {
      "get": {
        "operationId": "PaymentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      },
      "patch": {
        "operationId": "PaymentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      },
      "delete": {
        "operationId": "PaymentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{bookingId}/checkout": {
      "post": {
        "operationId": "PaymentController_checkout",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Sinh link thanh toán VNPAY (Sandbox)",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/ipn": {
      "post": {
        "operationId": "PaymentController_ipn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IpnPayloadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Webhook IPN từ cổng thanh toán",
        "tags": [
          "payments"
        ]
      }
    }
  },
  "info": {
    "title": "SportSpace API",
    "description": "Đặt sân thể thao + kết nối đồng đội",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "PLAYER",
              "MERCHANT",
              "ADMIN"
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "userId",
          "role"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateMatchDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string"
          },
          "slotsTotal": {
            "type": "number"
          },
          "skillLevel": {
            "type": "string"
          }
        },
        "required": [
          "bookingId",
          "slotsTotal"
        ]
      },
      "UpdateMatchDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string"
          },
          "slotsTotal": {
            "type": "number"
          },
          "skillLevel": {
            "type": "string"
          }
        }
      },
      "JoinMatchDto": {
        "type": "object",
        "properties": {}
      },
      "CreateVenueDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "address",
          "lat",
          "lng"
        ]
      },
      "UpdateVenueDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "courtId": {
            "type": "string"
          },
          "bookingDate": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        },
        "required": [
          "courtId",
          "bookingDate",
          "startTime",
          "endTime"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "courtId": {
            "type": "string"
          },
          "bookingDate": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        }
      },
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "bookingId",
          "provider",
          "amount"
        ]
      },
      "UpdatePaymentDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "returnUrl": {
            "type": "string"
          }
        }
      },
      "IpnPayloadDto": {
        "type": "object",
        "properties": {
          "transactionRef": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "signature": {
            "type": "string"
          }
        },
        "required": [
          "transactionRef",
          "amount",
          "signature"
        ]
      }
    }
  }
}